pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biomethod"
version = "0.1.0"
description = "Automated methods section generator for bioinformatics papers"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "BioMethod Contributors"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords = ["bioinformatics", "methods", "reproducibility", "citation", "workflow"]

dependencies = [
    "click>=8.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "python-docx>=0.8.11",
    "pandas>=1.5",
    "nbformat>=5.0",
    "openpyxl>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
biomethod = "biomethod.cli:main"

[project.urls]
Homepage = "https://github.com/biomethod/biomethod"
Documentation = "https://biomethod.readthedocs.io"
Repository = "https://github.com/biomethod/biomethod"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
biomethod = ["data/*.yaml", "generators/templates/*.j2"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
